+++ src/containers/CartContainer.jsx
@@ -2,13 +2,14 @@ import React from "react";
 import PropTypes from "prop-types";
 import { connect } from "react-redux";
 import { checkout } from "../actions";
-import { getTotal, getCartProducts } from "../reducers";
-import Cart from "../components/cart/Cart";
+import { getTotal, getCartProducts, getTotalItems } from "../reducers";
+import CartHeader from "../components/cart/CartHeader";
 
-const CartContainer = ({ products, total, checkout }) => (
-  <Cart
+const CartContainer = ({ products, total, checkout, totalItems }) => (
+  <CartHeader
     products={products}
     total={total}
+    totalItems={totalItems}
     onCheckoutClicked={() => checkout(products)}
   />
 );
@@ -31,6 +32,7 @@ CartContainer.propTypes = {
 const mapStateToProps = state => ({
   products: getCartProducts(state),
   total: getTotal(state),
+  totalItems: getTotalItems(state),
 });
 
 export default connect(mapStateToProps, { checkout })(CartContainer);
